+++ src/routes/api/size/[id]/+server.ts
@@ -7,46 +7,58 @@ import type { SizeDto } from '$lib/server/types/Dto';
 
 
 export const GET: RequestHandler = async ({ params }) => {
-	try {		
+	try {
+		const id = parseInt(params.id);
+
+		if (isNaN(id)) {
+			return json({ message: 'invalid ID '}, { status: 400 });
+		}
+
 		const result = await db.query.size.findFirst({
-			where: eq(size.id, parseInt(params.id))
+			where: eq(size.id, id)
 		});
 
 		if (!result) {
-			return json('Size not found', { status: 404 });
+			return json({ message: 'Size not found'}, { status: 404 });
 		}
 		
-		return json(result , { status: 200 });						
+		return json(result , { status: 200 });				
 		
 	}catch (error) {
 		console.error('Error fetching size:', error);
-		return json('Failed to fetch size', { status: 500 });
+		return json({ message: 'Failed to fetch size' }, { status: 500 });
 	}
-}
+};
 
 
 export const PUT: RequestHandler = async ({ request, params, cookies }) => {
 	try {
 		if (!cookies.get('session')) {
-			return json({ error: 'Unauthorized' }, { status: 401 });
+			return json({ message: 'Unauthorized' }, { status: 401 });
+		}
+
+		const id = parseInt(params.id);
+
+		if (isNaN(id)) {
+			return json({ message: 'invalid ID '}, { status: 400 });
 		}
 
 		const body: Partial<SizeDto> = await request.json();
 
 		const update = await db.update(size)
 								.set(body)
-								.where( eq(size.id, parseInt(params.id)) )
-								.returning();
+								.where( eq(size.id, id) )
+								.returning({ id: size.id, size: size.size, height: size.height });
 		
 		if (!update) {
-			return json('Size not found', { status: 404 });
+			return json({ message: 'Size not found' }, { status: 404 });
 		}
 
-		return json( update[0], { status: 200 });
+		return json( update, { status: 200 });
 
 	} catch (error) {
 		console.error('Error updating size:', error);
-		return json('Failed to updating size', { status: 500 });
+		return json({ message: 'Failed to updating size'}, { status: 500 });
 	}
 };
 
@@ -54,21 +66,27 @@ export const PUT: RequestHandler = async ({ request, params, cookies }) => {
 export const DELETE: RequestHandler = async ({ params, cookies }) => {
 	try {
 		if (!cookies.get('session')) {
-			return json('Unauthorized', { status: 401 });
+			return json({ message: 'Unauthorized'}, { status: 401 });
+		}
+
+		const id = parseInt(params.id);
+
+		if (isNaN(id)) {
+			return json({ message: 'invalid ID '}, { status: 400 });
 		}
 
 		const result = await db.delete(size)
-								.where(eq(size.id, parseInt(params.id)))
-								.returning()
+								.where( eq(size.id, id) )
+								.returning({ id: size.id, size: size.size, height: size.height});
 
 		
 		if (!result) {
-			return json('Size not found', { status: 404 });
+			return json({ message: 'Size not found'}, { status: 404 });
 		}
 
-		return new Response( null, { status: 204 });
+		return new Response(null, { status: 204 });
 	} catch (error) {
 		console.error('Error deleting size:', error);
-		return json('Failed to delete size', { status: 500 });
+		return json({ message: 'Failed to delete size' }, { status: 500 });
 	}
-}
+};
